Move WebStorageProvider to app module
