+++ app/src/main/java/com/duckduckgo/app/browser/mode/WebStorageProvider.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.browsermode.impl
+package com.duckduckgo.app.browser.mode
 
 import android.annotation.SuppressLint
 import android.webkit.WebStorage
@@ -23,18 +23,35 @@ import androidx.webkit.ProfileStore
 import com.duckduckgo.browsermode.api.BrowserMode
 import com.duckduckgo.browsermode.api.BrowserModeDataProvider
 import com.duckduckgo.browsermode.api.FireModeAvailability
+import com.duckduckgo.browsermode.api.profileName
 import com.duckduckgo.di.scopes.AppScope
+import com.squareup.anvil.annotations.ContributesTo
+import dagger.Binds
+import dagger.Module
 import dagger.SingleInstanceIn
 import javax.inject.Inject
 
+/**
+ * Resolves a per-mode [WebStorage]. Callers must invoke [forMode] on the main thread when
+ * `MULTI_PROFILE` is supported on the device — `ProfileStore` is tied to the WebView thread.
+ * Falls back to the default shared [WebStorage] when MultiProfile is unsupported.
+ */
 @SuppressLint("RequiresFeature")
 @SingleInstanceIn(AppScope::class)
 class WebStorageProvider @Inject constructor(
     private val fireModeAvailability: FireModeAvailability,
 ) : BrowserModeDataProvider<WebStorage> {
+
     @UiThread
     override fun forMode(mode: BrowserMode): WebStorage {
         if (!fireModeAvailability.isAvailable()) return WebStorage.getInstance()
         return ProfileStore.getInstance().getOrCreateProfile(mode.profileName).webStorage
     }
+}
+
+@ContributesTo(AppScope::class)
+@Module
+abstract class WebStorageProviderModule {
+    @Binds
+    abstract fun bindWebStorageProvider(impl: WebStorageProvider): BrowserModeDataProvider<WebStorage>
 }
